Add focus functionality
Created a new store reducer to handle the focus state of the weather items and hourly items. Also had to create
a separate util function for converting the unix timestamp. It is the only viable obtion to relate the weather item and hourly item. unixTimeConverter function is not suitable for this setup
since it was used only to display a readable data for the user, using it for focusing items created side effect that would cause not focusing items displaying data for 21:00 item.